misc/HtmlTestSuite/results/test-0415.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>The maximum value of scrollLeft shouldn't be affected by scrollbar even if it's wider than scroller</title>
<meta charset="utf-8"/>
<link href="https://drafts.csswg.org/cssom-view/#dom-element-scrollleft" rel="help"/>

</head><body><div id="scroller" style="overflow: scroll; width: 5px;height: 300px;">
    <div style="width:100px; height: 1px"/>
</div>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script>
    test(function() {
        scroller.scrollLeft = 1000;
        assert_equals(scroller.scrollLeft, 100);
    }, "Test the maxmium value of scrollLeft");
</script>
</body></html>
